fix CR #1317043

+++ sendimages/emailsettings.h
@@ -6,7 +6,7 @@
  * Date        : 2007-11-07
  * Description : e-mail settings container.
  *
- * Copyright (C) 2007-2014 by Gilles Caulier <caulier dot gilles at gmail dot com>
+ * Copyright (C) 2007-2015 by Gilles Caulier <caulier dot gilles at gmail dot com>
  * Copyright (C) 2010      by Andi Clemens <andi dot clemens at googlemail dot com>
  *
  * This program is free software; you can redistribute it
@@ -151,6 +151,7 @@ public:
                 return (*it).emailUrl;
             }
         }
+
         return KUrl();
     };
 
@@ -167,7 +168,7 @@ public:
 
     int              imageCompression;
 
-    int              attachmentLimitInMbytes;
+    qint64           attachmentLimitInMbytes;
 
     QString          tempPath;
     QString          tempFolderName;
